chore(registration-react): refactor rename api 'get-dataset' to 'datasets'
