+++ applications/registration-react/src/pages/api-registration-page/field-tagsinput-datasets/field-tagsinput-datasets.components.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import _ from 'lodash';
 import ReactTags from 'react-tag-autocomplete';
 import getTranslateText from '../../../lib/translateText';
-import { getDatasetByTitlePrefix } from '../../../api/get-dataset';
+import { getDatasetByTitlePrefix } from '../../../api/datasets';
 
 const addTagToInput = (updates, props) => {
   const { input } = props;
